feat: Add lazy loading for images in Home page
This commit adds lazy loading functionality for images in the Home page to improve page load performance. It updates the `ProductView` component to import the `useLanguage` hook and the `homePageLanguage` object from the `@/app/language/Lan-homepage` module. It also modifies the `HeroSection` component to import the `useLanguage` hook and the `homePageLanguage` object from the same module. Additionally, it updates the `HeroSection` component to use the `lan` variable for language localization and adds translations for the title and subheadings. Finally, it removes unnecessary code and fixes indentation issues in the `ProductView` component.