feat: Update image lazy loading and responsive styles
This commit updates the image lazy loading and responsive styles in multiple components. It modifies the `StatusFlow`, `OurService`, `CustomSlider`, `StatusFlowPoint`, `StatusCount`, `SecTitle`, `Navbar`, `ProductView`, `TrustedSection`, and `ContactHomeSection` components to include responsive classes for image heights and margins. It also adds the `lg` prefix to certain classes to apply styles only on larger screens. Additionally, it updates the `Lan-homepage` module to change the title of the `tabsSec` object to "Network of ZTL Offices". These changes improve the visual appearance and performance of the Home page.