chore: Update MegaNav component to fix active link bug
The MegaNav component was updated to fix a bug where the active link was not being correctly highlighted. The bug was caused by an incorrect comparison in the pathname check. This commit resolves the issue by updating the comparison to check for the correct pathname ("/Service"). Now, when the user is on the "/Service" page, the corresponding link in the MegaNav component will be correctly highlighted as active.